fix to do not use posthandler to handle failed tx sequence increment
